+++ CHANGELOG.md
@@ -15,6 +15,14 @@
 # <br><b>Changelog</b><br>
 
 
+<span id="v1-9-1" />
+
+## 21.11.2025 `v1.9.1`
+* Unified the module and class docstring styles throughout the whole library.
+* Moved the Protocol `ProgressUpdater` from the `console` module to the `types` module.
+* Added throttling to the `ProgressBar` update methods to impact the actual process' performance as little as possible.
+
+
 <span id="v1-9-0" />
 
 ## 21.11.2025 `v1.9.0` Big Update 🚀